microservicio_cargar_video/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==4.4.2
Flask==3.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
google-api-core==2.19.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
